Fix weird ES6 class problem with workaround
Apparently when we set a property on a constructor, an inheriting
class's constructor inherits that property. No, not from the
prototype, but from the constructor itself. This only happens if
the inheritor is an ES6 class, thus doesn't use the
constructor-inheritance pattern.

See: https://jsfiddle.net/5thnphkd/5/

To fix it this adds a search function that steps through the
prototype chain and checks whether a constructor above has the same
cached empty-value. If it does, that means that our empty-value is
actually fake, and inherited. Thus we have to ignore it.

This is so weird?!