Remove inefficient empty instance storage
The empty instance of a extendable (or sth inheriting it) have
to be stored, so that the same reference can be returned, thus
keeping the reference of the empty instances constant.

Before, the empty instance was stored in the constructor. But
since the constructor curiously inherits from the upper
constructor, there had to be search function, that checks
whether the reference actually stems from a parent. That was
really inefficient.

This was now replaced with a small store inside the extendable.js
that stores a prototype -> empty instance mapping.

It speeds up the perf test from 54μs to 45μs. Which is not far
away from pure Immutable.js's 42μs.